reprod/requirements.txt
ipykernel
ipython
jupyter_client
jupyter_core
matplotlib-inline
moviepy
numpy==1.26.4
pandas==2.2.2
requests
scipy==1.13.1
tqdm
autograd==1.6.2
feather-format
h5py
lifelines
matplotlib
nbclient
nbconvert
nbformat
nbsphinx
optuna==3.6.1
pykan==0.0.2
pyyaml
scikit-learn==1.5.0
sphinx
sphinx-autodoc-typehints
sphinx-rtd-theme
sphinxcontrib-applehelp
sphinxcontrib-devhelp
sphinxcontrib-htmlhelp
sphinxcontrib-jquery
sphinxcontrib-jsmath
sphinxcontrib-qthelp
sphinxcontrib-serializinghtml
sympy
torch==2.3.1
torchtuples==0.2.2
torchvision==0.18.1
py7zr
pysr
